refactor: move general grid-based systems to sokoban module
